api/requirements.txt
fastapi
uvicorn
pymongo
redis
docker
chromadb
requests
python-dotenv
openai
google-genai
